feat(tuning): time-budgeted bandit with a robust smoothed-curve score
Make the bandit reward a function of training time and a robust upper estimate
of the return curve, replacing the spiky, budget-biased plain max:

- Budget is now wall-clock training seconds (x eta each rung) instead of a
  rollout count, so slow-but-good arms are judged on what they achieve per unit
  time. The bandit's budget is a generic double resource the evaluator
  interprets.
- Arm score is the max of a rolling mean (window W) over the per-rollout return
  curve, rewarding a sustained plateau rather than a single lucky spike.

meta_train: --rung-seconds and --smoothing-window replace --rung-budget. Time
is measured over the step loop only (env startup / graph capture excluded as
fixed per-arm overhead); num_rollouts is lifted so wall-clock is the sole
limiter, which holds the LR ~constant within a probe.

+++ src/training/bandit.cc
@@ -173,11 +173,11 @@ std::vector<Arm> sample_arms(const Config &base, size_t count, uint64_t seed,
 }
 
 Bracket successive_halving(std::vector<Arm> arms, size_t eta,
-                           size_t rung_budget, const Evaluator &evaluate) {
+                           double rung_budget, const Evaluator &evaluate) {
   Bracket bracket;
   std::vector<Arm> alive = std::move(arms);
   size_t rung = 0;
-  size_t budget = rung_budget;
+  double budget = rung_budget;
   while (true) {
     RungResult result;
     result.rung = rung;
